move files to fit psr-4 declaration
